pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "firebot"
version = "2.0"
dependencies=[
    "lxml",
    "pandas",
    "requests",
    "python-dotenv"
]


